requirements.txt
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
Pillow>=10.0.0
